+++ assets/js/dashboard/nav-menu/query-periods/query-dates.test.tsx
@@ -10,7 +10,7 @@ import { QueryPeriodsPicker } from './query-periods-picker'
 const domain = 'picking-query-dates.test'
 const periodStorageKey = `period__${domain}`
 
-test('if no period is stored, loads with default value of "Last 30 days", all expected options are present', async () => {
+test('if no period is stored, loads with default value of "Last 28 days", all expected options are present', async () => {
   expect(localStorage.getItem(periodStorageKey)).toBe(null)
   render(<QueryPeriodsPicker />, {
     wrapper: (props) => (
@@ -18,7 +18,7 @@ test('if no period is stored, loads with default value of "Last 30 days", all ex
     )
   })
 
-  await userEvent.click(screen.getByText('Last 30 days'))
+  await userEvent.click(screen.getByText('Last 28 days'))
 
   expect(screen.getByTestId('datemenu')).toBeVisible()
   expect(screen.getAllByRole('link').map((el) => el.textContent)).toEqual(
@@ -27,9 +27,10 @@ test('if no period is stored, loads with default value of "Last 30 days", all ex
       ['Yesterday', 'E'],
       ['Realtime', 'R'],
       ['Last 7 Days', 'W'],
-      ['Last 30 Days', 'T'],
+      ['Last 28 Days', 'F'],
+      ['Last 90 Days', 'N'],
       ['Month to Date', 'M'],
-      ['Last Month', ''],
+      ['Last Month', 'P'],
       ['Year to Date', 'Y'],
       ['Last 12 Months', 'L'],
       ['All time', 'A'],
@@ -46,7 +47,7 @@ test('user can select a new period and its value is stored', async () => {
     )
   })
 
-  await userEvent.click(screen.getByText('Last 30 days'))
+  await userEvent.click(screen.getByText('Last 28 days'))
   expect(screen.getByTestId('datemenu')).toBeVisible()
   await userEvent.click(screen.getByText('All time'))
   expect(screen.queryByTestId('datemenu')).toBeNull()
@@ -162,7 +163,7 @@ test('going back resets the stored query period to previous value', async () => 
     }
   )
 
-  await userEvent.click(screen.getByText('Last 30 days'))
+  await userEvent.click(screen.getByText('Last 28 days'))
   await userEvent.click(screen.getByText('Year to Date'))
   expect(localStorage.getItem(periodStorageKey)).toBe('year')