+++ test/plausible_web/controllers/api/external_stats_controller/query_comparisons_test.exs
@@ -113,6 +113,128 @@ defmodule PlausibleWeb.Api.ExternalStatsController.QueryComparisonsTest do
            ]
   end
 
+  test "timeseries last 28d period compares the same period with and without match_day_of_week=true",
+       %{
+         conn: conn,
+         site: site
+       } do
+    today = ~D[2021-06-10]
+
+    make_request = fn match_day_of_week ->
+      conn
+      |> post("/api/v2/query-internal-test", %{
+        "site_id" => site.domain,
+        "metrics" => ["visitors"],
+        "date_range" => "28d",
+        "date" => Date.to_iso8601(today),
+        "dimensions" => ["time"],
+        "include" => %{
+          "time_labels" => true,
+          "comparisons" => %{
+            "mode" => "previous_period",
+            "match_day_of_week" => match_day_of_week
+          }
+        }
+      })
+      |> json_response(200)
+    end
+
+    assert %{"results" => results1} = make_request.(false)
+    assert %{"results" => results2} = make_request.(true)
+
+    assert results1 == results2
+
+    expected_first_date = today |> Date.shift(day: -28) |> Date.to_iso8601()
+    expected_last_date = today |> Date.shift(day: -1) |> Date.to_iso8601()
+    expected_comparison_first_date = today |> Date.shift(day: -56) |> Date.to_iso8601()
+    expected_comparison_last_date = today |> Date.shift(day: -29) |> Date.to_iso8601()
+
+    assert %{
+             "dimensions" => [actual_first_date],
+             "comparison" => %{
+               "dimensions" => [actual_comparison_first_date]
+             }
+           } = List.first(results1)
+
+    assert %{
+             "dimensions" => [actual_last_date],
+             "comparison" => %{
+               "dimensions" => [actual_comparison_last_date]
+             }
+           } = List.last(results1)
+
+    assert actual_first_date == expected_first_date
+    assert actual_last_date == expected_last_date
+    assert actual_comparison_first_date == expected_comparison_first_date
+    assert actual_comparison_last_date == expected_comparison_last_date
+  end
+
+  test "timeseries last 90d period in year_over_year comparison", %{
+    conn: conn,
+    site: site
+  } do
+    populate_stats(site, [
+      build(:pageview, timestamp: ~N[2021-04-01 00:00:00]),
+      build(:pageview, timestamp: ~N[2021-04-01 00:00:00]),
+      build(:pageview, timestamp: ~N[2021-04-05 00:00:00]),
+      build(:pageview, timestamp: ~N[2021-04-05 00:00:00]),
+      build(:pageview, timestamp: ~N[2021-06-29 00:00:00]),
+      build(:pageview, timestamp: ~N[2022-04-01 00:00:00]),
+      build(:pageview, timestamp: ~N[2022-04-05 00:00:00]),
+      build(:pageview, timestamp: ~N[2022-06-29 00:00:00]),
+      build(:pageview, timestamp: ~N[2022-06-30 00:00:00])
+    ])
+
+    conn =
+      post(conn, "/api/v2/query-internal-test", %{
+        "site_id" => site.domain,
+        "metrics" => ["visitors"],
+        "date_range" => "90d",
+        "date" => "2022-06-30",
+        "dimensions" => ["time:day"],
+        "include" => %{
+          "time_labels" => true,
+          "comparisons" => %{"mode" => "year_over_year"}
+        }
+      })
+
+    assert %{
+             "results" => results,
+             "meta" => %{"time_labels" => time_labels}
+           } = json_response(conn, 200)
+
+    assert "2022-04-01" = List.first(time_labels)
+    assert "2022-04-05" = Enum.at(time_labels, 4)
+    assert "2022-06-29" = List.last(time_labels)
+
+    assert %{
+             "dimensions" => ["2022-04-01"],
+             "metrics" => [1],
+             "comparison" => %{
+               "dimensions" => ["2021-04-01"],
+               "metrics" => [2]
+             }
+           } = Enum.find(results, &(&1["dimensions"] == ["2022-04-01"]))
+
+    assert %{
+             "dimensions" => ["2022-04-05"],
+             "metrics" => [1],
+             "comparison" => %{
+               "dimensions" => ["2021-04-05"],
+               "metrics" => [2]
+             }
+           } = Enum.find(results, &(&1["dimensions"] == ["2022-04-05"]))
+
+    assert %{
+             "dimensions" => ["2022-06-29"],
+             "metrics" => [1],
+             "comparison" => %{
+               "dimensions" => ["2021-06-29"],
+               "metrics" => [1]
+             }
+           } = Enum.find(results, &(&1["dimensions"] == ["2022-06-29"]))
+  end
+
   test "dimensional comparison with low limit", %{conn: conn, site: site} do
     populate_stats(site, [
       build(:pageview, browser: "Firefox", timestamp: ~N[2021-01-01 00:00:00]),